+++ src/module/nestAPI.module.ts
@@ -1,17 +1,19 @@
 import {Module} from "@nestjs/common";
 import {NestAPIController} from "./nestAPI.controller";
+import type {NestAPIGeneratorOptions} from "../script/generate";
 
-interface NestAPIModuleConfig {
-    path: string;
-}
+interface NestAPIModuleConfig extends NestAPIGeneratorOptions {}
 
 @Module({})
 export class NestAPIModule {
-    public static forRoot({path}: NestAPIModuleConfig) {
+    public static forRoot({appModule, globalPrefix}: NestAPIModuleConfig) {
         return {
             module: NestAPIModule,
             controllers: [NestAPIController],
-            providers: [{provide: "NEST_API_DIRECTORY", useValue: path}],
+            providers: [
+                {provide: "APP_MODULE", useValue: appModule},
+                {provide: "GLOBAL_PREFIX", useValue: globalPrefix},
+            ],
         };
     }
 }